refactor: move small components back to their parents files
