--- src/pages/TabDetails/CycleList/index.ts
@@ -1,3 +1,0 @@
-import CycleList from "./CycleList";
-
-export default CycleList;
