--- src/pages/TabDetails/index.ts
@@ -1,3 +1,0 @@
-import TabDetails from "./TabDetails";
-
-export default TabDetails;
