--- src/pages/TabHome/Calendar/interfaces.ts
@@ -1,3 +1,0 @@
-export interface CalendarProps {
-  setIsEditCalendar: (newIsOpen: boolean) => void;
-}
