--- src/pages/TabHome/DemoAlert.tsx
@@ -1,45 +1,0 @@
-import {
-  IonButton,
-  IonButtons,
-  IonCol,
-  IonModal,
-  IonToolbar,
-} from "@ionic/react";
-import { useState } from "react";
-import { useTranslation } from "react-i18next";
-
-const DemoAlert = () => {
-  const [isOpen, setIsOpen] = useState(true);
-  const { t } = useTranslation();
-
-  return (
-    <IonModal
-      id="alert-demo-modal"
-      isOpen={isOpen}
-    >
-      <div className="wrapper">
-        <h1>{t("This is just a demo")}</h1>
-        <p>
-          <span>{t("You can download the application ")}</span>
-          <a href="https://github.com/IraSoro/peri/releases/latest">
-            {t("here")}
-          </a>
-        </p>
-        <IonCol>
-          <IonToolbar>
-            <IonButtons slot="primary">
-              <IonButton
-                onClick={() => setIsOpen(false)}
-                color="dark-basic"
-              >
-                OK
-              </IonButton>
-            </IonButtons>
-          </IonToolbar>
-        </IonCol>
-      </div>
-    </IonModal>
-  );
-};
-
-export default DemoAlert;
