--- src/pages/TabHome/index.ts
@@ -1,3 +1,0 @@
-import TabHome from "./TabHome";
-
-export default TabHome;
