+++ src/tests/TabDetails.test.tsx
@@ -1,6 +1,6 @@
 import { it, expect } from "vitest";
 import { render } from "@testing-library/react";
-import TabDetails from "../pages/TabDetails";
+import TabDetails from "../pages/TabDetails/TabDetails";
 
 it("renders without crashing", () => {
   const { baseElement } = render(<TabDetails />);
